AI/requirements.txt
pandas==2.2.3
numpy==1.26.4
matplotlib==3.9.2
seaborn==0.12.2
python-chess==1.9.5
plotly==5.15.0
